+++ __init__.py
@@ -4,6 +4,15 @@
 # Changelog
 # 
 
+## [0.0.5] - 2021-30-10
+### Changed
+- World Control is now created in its own world named "WorldControl"
+
+### Added
+- Clean old setup when switching between advanced and basic
+- WorldControl has its own world, so we dont mess with user own setup worlds. Easier for cleaning
+- Easy switching operator, switch easy between basic and advanced
+
 ## [0.0.4] - 2021-30-10
 ### Changed
 - Moved panel to View category
@@ -34,7 +43,6 @@
 '''
 
 
-
 import os
 import bpy
 from bpy.types import Operator
@@ -42,10 +50,11 @@ from bpy.props import StringProperty
 from . import wc_3dv_panel
 # from bpy_extras.object_utils import AddObjectHelper
 
+
 bl_info = {
     "name": "World Control",
     "author": "Rombout Versluijs, Lech Sokolowski (Chocofur)",
-    "version": (0, 0, 4),
+    "version": (0, 0, 5),
     "blender": (2, 80, 0),
     "location": "World > Add Node > World Control",
     "description": "Adds a shader setup which allows more control when using HDR/EXR lighting. Based on Lech Sokolowski (Chocofur) video BCON19.",
@@ -55,10 +64,30 @@ bl_info = {
 }
 
 
+class WRD_OT_DeleteWorldControlBasic(Operator):
+    """Create a new World add World Control with basic settings"""
+    bl_idname = "world.delete_world_control"
+    bl_label = "Delete World Control"
+    bl_options = {'REGISTER', 'UNDO'}
 
-blendFile = "worldcontrol.blend"
+    def execute(self, context):
+        world = bpy.context.scene.world
+        nodes = world.node_tree.nodes
+        PrincipledWorldWrapper
+        principled_world = PrincipledWorldWrapper(world)
+        world_output = principled_world.node_out
+        for item in nodes:
+            if item != world_output:
+                # Skip over these when user adds WorldControl from different world 
+                # Otherwise we end with an empty fake custom node
+                if item.name != "Basic World" and item.name != "Advanced World":
+                    nodes.remove(item)
 
-# Source EasyHDRI
+        # self.report({'INFO'}, "Added Basic World Control")
+        return {'FINISHED'}
+
+
+# Source: LilySurfaceSrapper addon
 class PrincipledWorldWrapper:
     """This is a wrapper similar in use to PrincipledBSDFWrapper (located in
     bpy_extras.node_shader_utils) but for use with worlds. This is required to
@@ -74,7 +103,8 @@ class PrincipledWorldWrapper:
             elif self.node_out is None and n.type == "OUTPUT_WORLD":
                 self.node_out = n
 
-# Source LilySurfaceScraper
+
+# Source: LilySurfaceScraper
 # https://github.com/eliemichel/LilySurfaceScraper/blob/b1e7066f94b965bab90696e0be75faa44eb08c3e/blender/LilySurfaceScraper/CyclesWorldData.py
 def getGroundHdriNodeGroup(controlType):
     if ("WorldControl"+controlType) not in bpy.data.node_groups:
@@ -84,9 +114,7 @@ def getGroundHdriNodeGroup(controlType):
         filepath = blendfile + section + object
         directory = blendfile + section
         filename = object
-#        print(filepath)
-#        print(directory)
-#        print(filename)
+
         bpy.ops.wm.append(
             filepath=filepath,
             filename=filename,
@@ -95,21 +123,28 @@ def getGroundHdriNodeGroup(controlType):
 
 
 def addWorldControl(self, context,controlType):
-    #    world = bpy.data.worlds.new(name=self.name)
+    scn = bpy.context.scene
     # Create New world
-    # if bpy.data.worlds.find("World-Control") == -1:
-    #     world = bpy.data.worlds.new("World-Control")
-    # else:
-    #     world = bpy.data.worlds["World-Control"]
-    world = bpy.context.scene.world
+    if bpy.data.worlds.find("WorldControl") == -1:
+        world = bpy.data.worlds.new("WorldControl")
+    else:
+        world = bpy.data.worlds["WorldControl"]    
+
+    # org_world so we can delete fake custom node later
+    org_world = bpy.context.scene.world
+    scn.world = world
     world.use_nodes = True
 
     nodes = world.node_tree.nodes
     links = world.node_tree.links
+
     PrincipledWorldWrapper
     principled_world = PrincipledWorldWrapper(world)
     world_output = principled_world.node_out
     world_loc = world_output.location
+
+    # Clear old node setup
+    bpy.ops.world.delete_world_control()
 
     # Add World Control node group
     ground_hdri_node = nodes.new(type="ShaderNodeGroup")
@@ -117,7 +152,7 @@ def addWorldControl(self, context,controlType):
     ground_hdri_node.label = "WorldControl"+controlType
     ground_hdri_node.node_tree = getGroundHdriNodeGroup(controlType)
     ground_hdri_node.width = 185
-    ground_hdri_node.location = [world_loc.x - 200, world_loc.y]
+    ground_hdri_node.location = [world_loc.x - 235, world_loc.y]
 
     # Shade Node Types
     # https://docs.blender.org/api/current/bpy.types.html
@@ -139,6 +174,11 @@ def addWorldControl(self, context,controlType):
         links.new(mapping_node.outputs[0], texture_bg_node.inputs[0])
         links.new(texture_bg_node.outputs["Color"], ground_hdri_node.inputs["Background HDR"])
     links.new(ground_hdri_node.outputs[0], world_output.inputs["Surface"])
+
+    # Return orginal world so we clean fake custom node
+    scn.world = org_world
+    scn.world = world
+    return org_world
     
 
 class WRD_OT_AddWorldControlBasic(Operator):
@@ -148,8 +188,12 @@ class WRD_OT_AddWorldControlBasic(Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     def execute(self, context):
-        addWorldControl(self,context, "Basic")
-        bpy.context.scene.check_mode = "Basic"
+        bpy.context.scene.wc_check_mode = "Basic"
+        # We use org_world to be able to delete custo node
+        org_world = addWorldControl(self,context, "Basic")
+        nodes = org_world.node_tree.nodes
+        if nodes.find("Basic World") != -1:
+            nodes.remove(nodes["Basic World"])
         self.report({'INFO'}, "Added Basic World Control")
         return {'FINISHED'}
 
@@ -161,8 +205,12 @@ class WRD_OT_AddWorldControlAdvanced(Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     def execute(self, context):
-        addWorldControl(self,context, "Advanced")
-        bpy.context.scene.check_mode = "Advanced"
+        bpy.context.scene.wc_check_mode = "Advanced"
+        # We use org_world to be able to delete custo node
+        org_world = addWorldControl(self,context, "Advanced")
+        nodes = org_world.node_tree.nodes
+        if nodes.find("Advanced World") != -1:
+            nodes.remove(nodes["Advanced World"])
         self.report({'INFO'}, "Added Advanced World Control")
         return {'FINISHED'}
 
@@ -172,12 +220,11 @@ class ShaderNodeBasic(bpy.types.NodeCustomGroup):
     bl_label='Basic World'
 
     def init(self, context):
-        addWorldControl(self,context, "Basic")
-        world = bpy.context.scene.world
-        nodes = world.node_tree.nodes
+        # We use org_world to be able to delete custo node
+        org_world = addWorldControl(self,context, "Basic")
+        nodes = org_world.node_tree.nodes
         nodes.remove(nodes["Basic World"])
-        bpy.context.scene.check_mode = "Simple"
-
+        bpy.context.scene.wc_check_mode = "Basic"
 
 
 class ShaderNodeAdvanced(bpy.types.NodeCustomGroup):
@@ -185,16 +232,17 @@ class ShaderNodeAdvanced(bpy.types.NodeCustomGroup):
     bl_label='Advanced World'
 
     def init(self, context):
-        addWorldControl(self,context, "Advanced")
-        world = bpy.context.scene.world
-        nodes = world.node_tree.nodes
+        # We use org_world to be able to delete custo node
+        org_world = addWorldControl(self,context, "Advanced")
+        nodes = org_world.node_tree.nodes
         nodes.remove(nodes["Advanced World"])
-        bpy.context.scene.check_mode = "Advanced"
+        bpy.context.scene.wc_check_mode = "Advanced"
 
 
 from nodeitems_utils import NodeItem, register_node_categories, unregister_node_categories
 from nodeitems_builtins import ShaderNodeCategory
 from nodeitems_utils import NodeCategory, NodeItem
+
 
 class ShaderWorldNodeCategory(NodeCategory):
     @classmethod
@@ -202,22 +250,22 @@ class ShaderWorldNodeCategory(NodeCategory):
         # From Extra Material List addon > MeshLogic
         sdata = context.space_data
         return sdata.shader_type == 'WORLD'
-        # world = context.scene.world
-        # nodes = world.node_tree.nodes
-        # print(context.nodes)
-        # return nodes.nodes["World Output"]
-        # print(context.space_data.tree_type)
-        # return context.space_data.tree_type == 'ShaderNodeTree'
 
 
 node_categories = [ShaderWorldNodeCategory("WRL_NEW_CUSTOM","World Control",
                                         items=[NodeItem("ShaderNodeBasic",label ='Basic'),
                                                 NodeItem("ShaderNodeAdvanced",label ='Advanced')])] 
 
+# Call control mode > allows for easy switching
+# Source: https://blender.stackexchange.com/questions/70710/how-to-execute-an-operator-from-an-enumproperty
+def update_control_mode(self, context):
+    eval('bpy.ops.%s()' % self.wc_switch_control)
+
 
 classes = [
     ShaderNodeBasic,
     ShaderNodeAdvanced,
+    WRD_OT_DeleteWorldControlBasic,
     WRD_OT_AddWorldControlBasic,
     WRD_OT_AddWorldControlAdvanced
 ]
@@ -230,14 +278,22 @@ def register():
    
     register_node_categories("WCN_CUSTOM_NODES", node_categories)
     wc_3dv_panel.register()
-    bpy.types.Scene.check_mode = StringProperty(name="Mode",default="Basic",description="What control type are we using")#,update=wc_3dv_panel.check_world_check_mode)
-
+    bpy.types.Scene.wc_check_mode = StringProperty(name="Mode",default="Basic",description="What control type are we using")#,update=wc_3dv_panel.check_world_wc_check_mode)
+    bpy.types.Scene.wc_switch_control = bpy.props.EnumProperty(
+            name = "Mode",
+            description = "Switch contorl mode, switch between Basic and Advanced controls.",
+            items = [
+                ("world.add_world_control_basic", "Basic", "Basic mode allows for light, diffuse, background and glossy controls"),
+                ("world.add_world_control_advanced", "Advanced", "Basic mode allows for light, diffuse, background, reflection and glossy controls. Each items also has both strength and saturation. User can set different background for lighting and background."),              
+            ],
+            update=update_control_mode
+        )
 
 def unregister():
     for cls in classes:
         bpy.utils.unregister_class(cls)
     unregister_node_categories("WCN_CUSTOM_NODES")
-    del bpy.types.Scene.check_mode
+    del bpy.types.Scene.wc_check_mode
     wc_3dv_panel.unregister()